(#40): follow 컴포넌트 하나로 통합
