+++ src/pages/const/FOLLOW.tsx
@@ -14,16 +14,15 @@ import TenStraightIcon from '@/assets/follow/10시방향.svg?react';
 import EightStraightIcon from '@/assets/follow/8시방향.svg?react';
 import StartPoint from '@/assets/출발핀2.svg?react';
 import EndPoint from '@/assets/도착핀3.svg?react';
-import type { TurnType as PedestrianTurnType } from '../map/destination/types/pedestrianType';
-import type { TurnType as CarTurnType } from '../map/destination/types/carType';
 import { CarTunnelIcon } from '@/assets';
+import type { TurnType } from '../map/destination/types';
 
 interface FollowInfo {
   name: string;
   svg: React.ReactElement;
 }
 
-const baseFollowInfo = {
+export const followInfo: Record<TurnType, FollowInfo> = {
   11: {
     name: '직진',
     svg: <StraightIcon />,
@@ -56,6 +55,30 @@ const baseFollowInfo = {
     name: '4시 방향 우회전',
     svg: <FourStraightIcon />,
   },
+  121: {
+    name: '터널 진입',
+    svg: <CarTunnelIcon />,
+  },
+  125: {
+    name: '육교',
+    svg: <OverPassIcon />,
+  },
+  126: {
+    name: '지하보도',
+    svg: <UnderGroundIcon />,
+  },
+  127: {
+    name: '계단 진입',
+    svg: <StairsIcon />,
+  },
+  128: {
+    name: '경사로 진입',
+    svg: <RampIcon />,
+  },
+  129: {
+    name: '계단 + 경사로 진입',
+    svg: <StairsIcon />,
+  },
   184: {
     name: '경유지',
     svg: <StraightIcon />,
@@ -80,26 +103,7 @@ const baseFollowInfo = {
     name: '다섯 번째 경유지',
     svg: <StraightIcon />,
   },
-  125: {
-    name: '육교',
-    svg: <OverPassIcon />,
-  },
-  126: {
-    name: '지하보도',
-    svg: <UnderGroundIcon />,
-  },
-  127: {
-    name: '계단 진입',
-    svg: <StairsIcon />,
-  },
-  128: {
-    name: '경사로 진입',
-    svg: <RampIcon />,
-  },
-  129: {
-    name: '계단 + 경사로 진입',
-    svg: <StairsIcon />,
-  },
+
   200: {
     name: '출발지',
     svg: <StartPoint />,
@@ -139,17 +143,5 @@ const baseFollowInfo = {
   218: {
     name: '엘리베이터',
     svg: <ElevatorIcon />,
-  },
-};
-
-export const pedestrianFollowInfo: Record<PedestrianTurnType, FollowInfo> = {
-  ...baseFollowInfo,
-};
-
-export const carFollowInfo: Record<CarTurnType, FollowInfo> = {
-  ...baseFollowInfo,
-  121: {
-    name: '터널 진입',
-    svg: <CarTunnelIcon />,
   },
 };
