+++ src/pages/map/destination/types/index.ts
@@ -1,14 +1,12 @@
 import type {
   CarFeatures,
   SearchOption as CarSearchOptions,
-  TurnType as CarTurnType,
   ResponseRoadType,
   FacilityType as CarFacilityType,
 } from './carType';
 import type {
   PedestrianFeatures,
   SearchOption as PedestrianSearchOptions,
-  TurnType as PedestrianTurnType,
   FacilityType as PedestrianFacilityType,
 } from './pedestrianType';
 
@@ -22,7 +20,6 @@ export type {
   DetailPosFlag as CarDetailPosFlag,
   TotalValue as CarTotalValue,
   Traffic as CarTraffic,
-  TurnType as CarTurnType,
   PointType as CarPointType,
   PointGeometry as CarPointGeometry,
   LineStringGeometry as CarLineStringGeometry,
@@ -43,7 +40,6 @@ export type {
   SearchOption as PedestrianSearchOption,
   SortType as PedestrianSortType,
   RoadType as PedestrianRoadType,
-  TurnType as PedestrianTurnType,
   FacilityType as PedestrianFacilityType,
   PointProperties as PedestrianPointProperties,
   LineStringProperties as PedestrianLineStringProperties,
@@ -51,6 +47,38 @@ export type {
   PedestrianResponse,
   PedestrianRequestBody,
 } from './pedestrianType';
+
+export type TurnType =
+  | 11 //직진
+  | 12 //좌회전
+  | 13 //우회전
+  | 14 //유턴
+  | 16 //8시 방향 좌회전
+  | 17 //10시 방향 좌회전
+  | 18 //2시 방향 우회전
+  | 19 //4시 방향 우회전
+  | 184 //경유지
+  | 185 //첫 번째 경유지
+  | 186 //두 번째 경유지
+  | 187 //세 번째 경유지
+  | 188 //네 번째 경유지
+  | 189 //다섯 번째 경유지
+  | 121 //터널진입
+  | 125 //육교
+  | 126 //지하보도
+  | 127 //계단 진입
+  | 128 // 경사로 진입
+  | 129 //계단 + 경사로 진입
+  | 200 // 출발지
+  | 201 // 목적지
+  | 211 //횡단보도
+  | 212 //좌측 횡산보도
+  | 213 //우측 횡단보도
+  | 214 // 8시 방향 횡단보도
+  | 215 // 10시 방향 횡단보도
+  | 216 // 2시 방향 횡단보도
+  | 217 // 4시 방향 횡단보도
+  | 218; // 엘리베이터
 
 export type TransportationType =
   | 'pedestrian'
@@ -72,7 +100,7 @@ export interface FollowBase {
 export interface PedestrianFollowFeature extends FollowBase {
   totalTime?: number;
   totalDistance?: number;
-  turnType: PedestrianTurnType;
+  turnType: TurnType;
   facilityType?: PedestrianFacilityType;
 }
 
@@ -80,7 +108,7 @@ export interface CarFollowFeature extends FollowBase {
   totalTime?: number;
   totalDistance?: number;
   taxiFare?: number;
-  turnType: CarTurnType;
+  turnType: TurnType;
   roadType?: ResponseRoadType;
   facilityType?: CarFacilityType;
 }
